+++ tests/test_settings.rs
@@ -28,32 +28,6 @@ fn test_simple() {
 
 #[cfg(feature = "yaml")]
 #[test]
-#[allow(deprecated)]
-fn test_bound_to_thread() {
-    let mut map = std::collections::HashMap::new();
-    map.insert("a", "first value");
-    map.insert("b", "second value");
-    map.insert("c", "third value");
-    map.insert("d", "fourth value");
-
-    let mut settings = Settings::new();
-    settings.set_sort_maps(true);
-    settings.bind_to_thread();
-    assert_yaml_snapshot!(&map, @r###"
-    ---
-    a: first value
-    b: second value
-    c: third value
-    d: fourth value
-    "###);
-
-    // put defaults back
-    let settings = Settings::new();
-    settings.bind_to_thread();
-}
-
-#[cfg(feature = "yaml")]
-#[test]
 fn test_bound_to_scope() {
     let mut map = std::collections::HashMap::new();
     map.insert("a", "first value");